Improve internal decomposition
